// FRONTEND/chatbot-admin-panel/src/services/api/schemas/databaseConnectionCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DatabaseConnectionCreateDescription } from './databaseConnectionCreateDescription';
import type { DBType } from './dBType';
import type { DatabaseConnectionCreateExtraParams } from './databaseConnectionCreateExtraParams';
import type { DatabaseConnectionCreatePassword } from './databaseConnectionCreatePassword';

export interface DatabaseConnectionCreate {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: DatabaseConnectionCreateDescription;
  db_type: DBType;
  host: string;
  /**
   */
  port: number;
  database_name: string;
  username: string;
  extra_params?: DatabaseConnectionCreateExtraParams;
  password?: DatabaseConnectionCreatePassword;
}