// FRONTEND/chatbot-admin-panel/src/services/api/schemas/databaseConnectionResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DatabaseConnectionResponseDescription } from './databaseConnectionResponseDescription';
import type { DBType } from './dBType';
import type { DatabaseConnectionResponseExtraParams } from './databaseConnectionResponseExtraParams';

export interface DatabaseConnectionResponse {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: DatabaseConnectionResponseDescription;
  db_type: DBType;
  host: string;
  /**
   */
  port: number;
  database_name: string;
  username: string;
  extra_params?: DatabaseConnectionResponseExtraParams;
  id: number;
  created_at: string;
  updated_at: string;
}