// FRONTEND/chatbot-admin-panel/src/services/api/schemas/dBTypeSchema.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un chatbot avanzado con IA, configurable y con gestión de permisos.
 * OpenAPI spec version: 0.1.0
 */

export type DBTypeSchema = typeof DBTypeSchema[keyof typeof DBTypeSchema];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DBTypeSchema = {
  POSTGRESQL: 'POSTGRESQL',
  SQLSERVER: 'SQLSERVER',
  MYSQL: 'MYSQL',
  ORACLE: 'ORACLE',
} as const;